+++ .gitignore
@@ -1,1 +1,1 @@
-*.exe
+/.vscode/
