+++ src/functions.cpp
